refactor(:train:): move utility method for custom params into request
AFFECTS PACKAGES:
@esri/arcgis-rest-feature-service
@esri/arcgis-rest-request
